created a helper files for redux folder and moved some logic from action creators to it
