+++ src/redux/actionCreators.js
@@ -1,5 +1,6 @@
 import * as types from './actionTypes';
 import axios from 'axios';
+import { mapPromise } from './helpers'
 
 const url = 'https://where2code.herokuapp.com/api';
 
@@ -8,12 +9,10 @@ export const locationSuccess = locationList => ({
   type: types.FETCH_LOCATIONS_SUCCESS,
   payload: locationList.data
 });
-
 export const locationFailure = error => ({
   type: types.FETCH_LOCATIONS_FAILURE,
   payload: error
 });
-
 export const locationLoads = currentLocation => async dispatch => {
   dispatch({ type: types.LOADING_LOCATIONS });
   try {
@@ -31,26 +30,14 @@ export const mapsSucces = (mapsObj, geolocation) => ({
   type: types.FETCH_MAP_API_SUCCESS,
   payload: { mapsObj, geolocation }
 });
-
 export const mapsFailure = error => ({
   type: types.FETCH_MAP_API_FAILURE,
   payload: error
 });
-
 export const mapsLoading = geolocation => async dispatch => {
   dispatch({ type: types.LOADING_MAP_API });
   try {
-    const mapPromise = new Promise((resolve, reject) => {
-      window.initGoogleMapPromise = () => {
-        resolve(window.google);
-        delete window.initGoogleMapPromise;
-      };
-      const script = document.createElement('script');
-      const GOOGLE_API_KEY = process.env.REACT_APP_GOOGLE_API_KEY;
-      script.async = true;
-      script.src = `https://maps.googleapis.com/maps/api/js?key=${GOOGLE_API_KEY}&callback=initGoogleMapPromise`;
-      document.body.appendChild(script);
-    });
+    // We import the Promise from helpers here
     Promise.all([mapPromise]).then(value => {
       dispatch(mapsSucces(value[0].maps, geolocation));
     });
@@ -58,27 +45,22 @@ export const mapsLoading = geolocation => async dispatch => {
     dispatch(mapsFailure(error.message));
   }
 };
-
 export const setGeolocationTrue = () => ({
   type: types.SET_GEOLOCATION_TRUE
 })
-
 export const setGeolocationFalse = () => ({
   type: types.SET_GEOLOCATION_FALSE
 })
-
 
 // ACTIONS FOR SINGLE LOCATION REDUCER
 export const singleLocSuccess = locationList => ({
   type: types.FETCH_SINGLE_LOCATIONS_SUCCESS,
   payload: locationList
 });
-
 export const singleLocFailure = error => ({
   type: types.FETCH_SINGLE_LOCATIONS_FAILURE,
   payload: error
 });
-
 export const fetchSingleLocation = locId => async dispatch => {
   dispatch({ type: types.LOADING_SINGLE_LOCATION });
   try {
